주석 첨가 및 xml 수정
controller
- 주석 첨가
- 버그 수정 (질문 작성 시)
- 예외처리 수정 (커스터마이징 한 에러페이지로 넘어갈 수 있도록)

xml
-resultType hashmap을 각자의 도메인객체로 수정

+++ src/main/java/com/learning/controller/BoardController.java
@@ -53,9 +53,12 @@ public class BoardController {
 	//질문 게시판 리스트
 	@GetMapping("/list")
 	public String boardList(Criteria criteria, Model model) {
-
+		
+		//페이징의 기준이 되는 객체 criteria를 가지고 질문 게시글 리스트 불러오기
 		List<Board> boardList = boardService.getBoardList(criteria);
+		// 질문 게시글의 총 갯수를 구해서
 		int total = boardService.getBoardTotal();
+		//페이지메이커로 페이지네이션
 		PageMaker pageMaker = new PageMaker(criteria, total);
 
 		model.addAttribute("pmk", pageMaker);
@@ -67,8 +70,10 @@ public class BoardController {
 	@GetMapping("/listSearch")
 	public String boardSearchedList(@RequestParam("option") String option,
 			@RequestParam("search") String search, Criteria criteria, Model model) {
-		 
+		
 		int total = 0;
+		
+		//검색 옵션에 따라 검색 결과가 다르게 (과목명, 문제번호, 게시글제목, 글쓴이)
 		List<Board> boardList = null;
 			if (option.equals("subject")) {
 				boardList = boardService.getBoardListBySubject(criteria, search);
@@ -98,7 +103,10 @@ public class BoardController {
 	public String boardWrite(@RequestParam(value = "question_no", required = false) Long question_no,
 			@ModelAttribute("boardForm") BoardForm boardForm, Principal principal, Model model) {
 
+		//작성자 id
 		String username = principal.getName();
+		
+		//문제와 연동된 질문글일 시 문제 정보와 과목이름을 가져옴
 		if (question_no != null) {
 			Question question = questionService.getQuestionByQNo(question_no);
 			String subject_name = subjectService.getSubjectNameByQuestion(question.getSubject_code());
@@ -114,12 +122,26 @@ public class BoardController {
 	//질문 게시글 작성
 	@PostMapping("/write")
 	public String boardWrite(@Valid BoardForm boardForm, BindingResult result, Model model) {
-
+		
+		//유효성 검사를 통과하지 못했을 시 form으로 되돌아가기
 		if (result.hasErrors()) {
+			
+			//문제와 연동 된 게시글일 경우
+			if(boardForm.getQuestion_no()!=null) {
+				Question question = questionService.getQuestionByQNo(boardForm.getQuestion_no());
+				String subject_name = subjectService.getSubjectNameByQuestion(question.getSubject_code());
+				
+				model.addAttribute("question", question);
+				model.addAttribute("subject_name", subject_name);
+			}
+			//form에 표시할 username
+			model.addAttribute("username",boardForm.getUsername());
 			return "board/write";
 		}
-
+		
+		//질문게시글 DB에 들어갈 과목코드 가져오기
 		String subject_code = subjectService.getSubjectCodeByQNo(boardForm.getQuestion_no());
+		//질문게시글 DB에 들어갈 글쓴이정보 (user_no) 가져오기
 		Long user_no = userService.findUserNoById(boardForm.getUsername());
 		boardService.writeBoard(user_no, subject_code, boardForm);
 
@@ -151,7 +173,7 @@ public class BoardController {
 		// 현재 로그인 되어있는 유저의 id 전달
 		String loginUser = principal.getName();
 		
-		// 최초 표시할 댓글리스트 전탈
+		// 최초 표시할 댓글리스트 전달
 		List<Answer> answerList = answerService.getAnswerList(board_no);
 		
 		model.addAttribute("answerList",answerList);
@@ -166,21 +188,16 @@ public class BoardController {
 		
 		Board board = boardService.getBoard(board_no);
 		
-		if(board==null) {
-			model.addAttribute("message","해당 게시글을 찾을 수 없습니다.");
-			model.addAttribute("url","/board/list");
-			return "alert";
-		}
-		
+		//게시글 수정 권한이 없는 유저가 주소창에 url로 직접 접근 시도 시 error 페이지 (alert 후 뒤로가기)
 		if (!board.getUsername().equals(principal.getName()) && !principal.getName().equals("admin")) {
 			return "error/403";
 		}
 		
+		//문제와 연동된 질문글일 경우 문제 정보 가져와서 수정 페이지에 보여주기
 		if(board.getQuestion_no()!=null) {
 			Question question = questionService.getQuestionByQNo(board.getQuestion_no());
 			model.addAttribute("question",question);
 		}
-		
 		
 		model.addAttribute("board",board);
 		
@@ -191,14 +208,14 @@ public class BoardController {
 	@PostMapping("/modify/{board_no}")
 	public String boardModify(@PathVariable("board_no")Long board_no, @Valid BoardForm boardForm, BindingResult result, Model model) {
 		
+		//유효성 검사를 통과하지 못했을 때
 		if(result.hasErrors()) {
 			Board board = boardService.getBoard(board_no);
-			
+			//문제와 연동 된 질문글일 경우 문제 정보 가져오기
 			if(board.getQuestion_no()!=null) {
 				Question question = questionService.getQuestionByQNo(board.getQuestion_no());
 				model.addAttribute("question",question);
 			}
-			
 			model.addAttribute("board",board);
 			return "board/modify";
 		}
@@ -212,13 +229,18 @@ public class BoardController {
 	
 	//게시글 삭제
 	@GetMapping("/delete/{board_no}")
-	public String boardDelete(@PathVariable("board_no")Long board_no, Model model) {
+	public String boardDelete(@PathVariable("board_no")Long board_no, Principal principal, Model model) {
 		
 		Board board = boardService.getBoard(board_no);
+		//주소창에 url로 직접 접근 시 없는 게시글일 경우 alert
 		if(board == null) {
 			model.addAttribute("message","해당 게시글을 찾을 수 없습니다.");
 			model.addAttribute("url","/board/list");
 			return "alert";
+		}
+		//삭제 권한이 없는 유저가 주소창으로 직접 접근 시 에러페이지
+		if(!principal.getName().equals(board.getUsername()) && !principal.getName().equals("admin")){
+			return "error/403";
 		}
 		
 		boardService.deleteBoard(board_no);
@@ -228,13 +250,16 @@ public class BoardController {
 	}
 	
 	
-	//공지글 리스트
+	//공지글 리스트 (비회원도 가능)
 	@PreAuthorize("permitAll()")
 	@GetMapping("/notice")
 	public String noticeList(Criteria criteria, Model model) {
 		
+		//페이징 기준 객체인 criteria를 가지고 공지글 리스트를 가져옴
 		List<Board> noticeList = boardService.getNoticeList(criteria);
+		//공지글의 총 갯수를 구해서
 		int total = boardService.getNoticeTotal();
+		//criteria를 가지고 페이지네이션
 		PageMaker pageMaker = new PageMaker(criteria, total);
 
 		model.addAttribute("pmk", pageMaker);
@@ -243,13 +268,16 @@ public class BoardController {
 		return "board/notice";
 	}
 	
-	//공지글 검색
+	//공지글 검색 (비회원도 가능)
 	@PreAuthorize("permitAll()")
 	@GetMapping("/noticeSearch")
 	public String noticeSearch(@RequestParam("search")String search,Criteria criteria, Model model) {
 		
+		//페이징 기준 객체인 criteria와 검색어를 가지고 검색된 공지글 리스트를 가져옴
 		List<Board> noticeList = boardService.getNoticeByTitle(criteria, search);
+		//검색 된 공지의 총 갯수를 구해서
 		int total = boardService.getNoticeTotalBySearch(search);
+		//페이지네이션
 		PageMaker pageMaker = new PageMaker(criteria, total);
 
 		model.addAttribute("pmk", pageMaker);
@@ -263,11 +291,9 @@ public class BoardController {
 	@GetMapping("/notice/{board_no}")
 	public String noticeRead(@PathVariable("board_no")Long board_no, Principal principal ,Model model) {
 		Board board = boardService.getBoard(board_no);
-		if(board==null) {
-			model.addAttribute("message","해당 게시글을 찾을 수 없습니다.");
-			model.addAttribute("url","/board/list");
-			return "alert";
-		}
+		
+		//로그인 한 회원 id를 가져옴 (공지를 수정,삭제 할 권한이 있는 지 판단하기 위함)
+		//조건문을 붙이지 않으면 null오류 발생
 		if(principal!=null) {
 			model.addAttribute("loginUser",principal.getName());
 		}
@@ -282,6 +308,7 @@ public class BoardController {
 	@GetMapping("/noticeWrite")
 	public String noticeWrite(@ModelAttribute("boardForm")BoardForm boardForm, Principal principal, Model model) {
 		
+		//공지를 작성하는 사람의 id
 		String username = principal.getName();
 		model.addAttribute("username",username);
 		return "board/noticeWrite";
@@ -296,6 +323,7 @@ public class BoardController {
 			model.addAttribute("username",principal.getName());
 			return "board/noticeWrite";
 		}
+		//게시글 DB에 저장될 user_no 가져오기
 		Long user_no = userService.findUserNoById(principal.getName());
 		boardService.writeNotice(user_no, boardForm);
 		model.addAttribute("message","공지 작성 완료");
@@ -309,10 +337,12 @@ public class BoardController {
 	public String noticeModify(@PathVariable("board_no")Long board_no, @ModelAttribute("boardForm")BoardForm boardForm, Principal principal, Model model) {
 		
 		Board board = boardService.getBoard(board_no);
+		// 주소창으로 직접 입력 시 없는 게시글 번호일 경우
 		if (board==null) {
-			return "error/403";
+			return "error/500";
 		}
 		
+		// 권한이 없는 유저가 직접 주소창에 입력 할 경우 (admin은 모든 공지를 수정/삭제 가능하지만, 선생님 계정은 본인이 쓴 공지만 수정/삭제 가능)
 		if (!board.getUsername().equals(principal.getName()) && !principal.getName().equals("admin")) {
 			return "error/403";
 		}
@@ -326,6 +356,7 @@ public class BoardController {
 	@PostMapping("/noticeModify/{board_no}")
 	public String noticeModify(@PathVariable("board_no")Long board_no, @Valid BoardForm boardForm, BindingResult result, Model model) {
 		
+		//유효성 검사를 통과 못했을 시
 		if(result.hasErrors()) {
 			Board board = boardService.getBoard(board_no);
 			model.addAttribute("board",board);
@@ -344,10 +375,12 @@ public class BoardController {
 	public String noticeDelete(@PathVariable("board_no")Long board_no, Principal principal, Model model) {
 		
 		Board board = boardService.getBoard(board_no);
-		if (board==null) {
-			return "error/403";
-		}
 		
+		//주소창에 직접 url을 입력했을 때 해당 번호의 게시글이 없는 경우
+		if (board==null) {
+			return "error/500";
+		}
+		//권한이 없는 유저가 주소창으로 직접 입력해서 접근 한 경우
 		if (!board.getUsername().equals(principal.getName()) && !principal.getName().equals("admin")) {
 			return "error/403";
 		}
